Move all geometry code into the separate folder
