+++ Renderer/Renderer/RayCasting/Geometry/SceneGeometry.cpp
@@ -1,6 +1,6 @@
 #include <RendererStdAfx.h>
 
-#include <Renderer/RayCasting/SceneGeometry.h>
+#include <Renderer/RayCasting/Geometry/SceneGeometry.h>
 
 namespace GLEngine::Renderer {
