+++ Renderer/Renderer/RayCasting/Light/RayPointLight.h
@@ -12,7 +12,7 @@ public:
 	C_PointLight(const glm::vec3& position, const glm::vec3& intenstiy);
 	[[nodiscard]] glm::vec3 SampleLi(const C_RayIntersection& intersection, I_Sampler* rnd, S_VisibilityTester& vis, float* pdf) const override;
 	[[nodiscard]] glm::vec3 Le() const override;
-	[[nodiscard]] float	  PdfLe() const override;
+	[[nodiscard]] float		PdfLe() const override;
 
 private:
 	glm::vec3 m_Position;
